重构了Python接口结构
